polymarket-agent/requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
bcrypt==4.2.0
bitarray==2.9.2
build==1.2.1
cachetools==5.4.0
certifi==2024.7.4
cfgv==3.4.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.5.5
ckzg==1.0.2
click==8.1.7
coloredlogs==15.0.1
cytoolz==0.12.3
dataclasses-json==0.6.7
Deprecated==1.2.14
devtools==0.12.2
distlib==0.3.8
distro==1.9.0
dnspython==2.6.1
eip712-structs==1.1.0
email_validator==2.2.0
eth-account==0.13.1
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==2.1.0
eth-typing==4.4.0
eth-utils==4.1.1
eth_abi==5.1.0
exceptiongroup==1.2.2
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.15.4
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.6.1
google-auth==2.32.0
googleapis-common-protos==1.63.2
grpcio==1.65.2
h11==0.14.0
hexbytes==1.2.1
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.5
humanfriendly==10.0
identify==2.6.0
idna==3.7
importlib_metadata==8.0.0
importlib_resources==6.4.0
iniconfig==2.0.0
Jinja2==3.1.4
jq==1.7.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kubernetes==30.1.0
langchain==0.2.11
langchain-chroma==0.1.2
langchain-community==0.2.10
langchain-core==0.2.26
langchain-openai==0.1.19
langchain-text-splitters==0.2.2
langchainhub==0.1.20
langgraph==0.1.17
langsmith==0.1.94
lru-dict==1.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
newsapi-python==0.2.7
nodeenv==1.9.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.18.1
openai==1.37.1
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-instrumentation==0.47b0
opentelemetry-instrumentation-asgi==0.47b0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-util-http==0.47b0
orjson==3.10.6
overrides==7.7.0
packaging==24.1
pandas>=2.0.0
parsimonious==0.10.0
platformdirs==4.2.2
plotly>=5.17.0
pluggy==1.5.0
poly_eip712_structs==0.0.1
posthog==3.5.0
pre-commit==3.8.0
protobuf==4.25.4
py_clob_client==0.17.5
py_order_utils==0.3.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyPika==0.48.9
pyproject_hooks==1.1.0
pysha3==1.0.2
pytest==8.3.2
pytest-asyncio>=0.21.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pyunormalize==15.1.0
PyYAML==6.0.1
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rlp==4.0.1
rpds-py==0.19.1
rsa==4.9
scheduler==0.8.7
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.37.2
streamlit>=1.28.0
sympy==1.13.1
tavily-python==0.3.5
tenacity==8.5.0
tiktoken==0.7.0
tokenizers==0.19.1
tomli==2.0.1
toolz==0.12.1
tqdm==4.66.4
typeguard==4.3.0
typer==0.12.3
types-requests==2.32.0.20240712
typing-inspect==0.9.0
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.3
uvloop==0.19.0
virtualenv==20.26.3
watchfiles==0.22.0
web3==6.11.0
websocket-client==1.8.0
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.19.2
